src/requirements.txt
aiogram==3.15.0
fastapi==0.115.6
pydantic==2.10.4
pymongo==4.10.1
python-dotenv==1.0.1
uvicorn==0.34.0
